# spec/04-openapi.yaml
openapi: 3.1.0
info: { title: WALeads API, version: 0.1.0 }
servers: [{ url: http://localhost:3000 }]
paths:
  /health:
    get: { summary: Health check, responses: { '200': { description: Disponible } } }
  /api/v1/webhooks/whatsapp:
    get:
      summary: Handshake de Meta
      parameters:
        - { in: query, name: hub.mode, required: true, schema: { type: string } }
        - { in: query, name: hub.verify_token, required: true, schema: { type: string } }
        - { in: query, name: hub.challenge, required: true, schema: { type: string } }
      responses: { '200': { description: Verificado }, '403': { description: Token inválido } }
    post:
      summary: Recepción de mensajes
      parameters: [{ in: header, name: x-hub-signature-256, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Recibido }, '401': { description: Firma inválida } }
  /api/v1/leads:
    get:
      summary: Listar leads
      security: [{ bearerAuth: [] }]
      parameters: [{ in: query, name: search, schema: { type: string } }, { in: query, name: page, schema: { type: integer, minimum: 1 } }, { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100 } }]
      responses: { '200': { description: Lista paginada }, '401': { description: JWT ausente o inválido } }
  /api/v1/pedidos:
    get:
      summary: Listar pedidos
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Lista paginada }, '401': { description: JWT ausente o inválido } }
    post:
      summary: Crear pedido
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NewOrder' } } } }
      responses: { '201': { description: Creado }, '401': { description: JWT ausente o inválido }, '404': { description: Lead inexistente } }
  /api/v1/pedidos/{id}/estado:
    patch:
      summary: Cambiar estado
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [estado], properties: { estado: { $ref: '#/components/schemas/OrderState' } } } } } }
      responses: { '200': { description: Actualizado }, '401': { description: JWT ausente o inválido }, '409': { description: Transición inválida } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OrderState: { type: string, enum: [PENDIENTE, EN_PREPARACION, DESPACHADO, ENTREGADO, CANCELADO] }
    NewOrder:
      type: object
      required: [leadId, items, direccionEntrega]
      properties:
        leadId: { type: string }
        direccionEntrega: { type: string }
        moneda: { type: string, enum: [ARS, USD], default: ARS }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productoId, descripcion, cantidad, precioUnitario]
            properties: { productoId: { type: string }, descripcion: { type: string }, cantidad: { type: integer, minimum: 1 }, precioUnitario: { type: number, minimum: 0 } }